Move the student progress logic to the internal namespace
