+++ includes/class-sensei.php
@@ -6,12 +6,12 @@ use Sensei\Quiz_Submission\Grade\Repositories\Grade_Repository_Factory;
 use Sensei\Quiz_Submission\Grade\Repositories\Grade_Repository_Interface;
 use Sensei\Quiz_Submission\Submission\Repositories\Submission_Repository_Factory;
 use Sensei\Quiz_Submission\Submission\Repositories\Submission_Repository_Interface;
-use Sensei\Student_Progress\Course_Progress\Repositories\Course_Progress_Repository_Factory;
-use Sensei\Student_Progress\Course_Progress\Repositories\Course_Progress_Repository_Interface;
-use Sensei\Student_Progress\Lesson_Progress\Repositories\Lesson_Progress_Repository_Factory;
-use Sensei\Student_Progress\Lesson_Progress\Repositories\Lesson_Progress_Repository_Interface;
-use Sensei\Student_Progress\Quiz_Progress\Repositories\Quiz_Progress_Repository_Factory;
-use Sensei\Student_Progress\Quiz_Progress\Repositories\Quiz_Progress_Repository_Interface;
+use Sensei\Internal\Student_Progress\Course_Progress\Repositories\Course_Progress_Repository_Factory;
+use Sensei\Internal\Student_Progress\Course_Progress\Repositories\Course_Progress_Repository_Interface;
+use Sensei\Internal\Student_Progress\Lesson_Progress\Repositories\Lesson_Progress_Repository_Factory;
+use Sensei\Internal\Student_Progress\Lesson_Progress\Repositories\Lesson_Progress_Repository_Interface;
+use Sensei\Internal\Student_Progress\Quiz_Progress\Repositories\Quiz_Progress_Repository_Factory;
+use Sensei\Internal\Student_Progress\Quiz_Progress\Repositories\Quiz_Progress_Repository_Interface;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
