+++ includes/internal/student-progress/course-progress/repositories/class-course-progress-repository-interface.php
@@ -1,0 +1,55 @@
+<?php
+/**
+ * File containing the Course_Progress_Repository_Interface.
+ *
+ * @package sensei
+ */
+
+namespace Sensei\Internal\Student_Progress\Course_Progress\Repositories;
+
+use Sensei\Internal\Student_Progress\Course_Progress\Models\Course_Progress;
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+/**
+ * Interface Course_Progress_Repository_Interface.
+ *
+ * @since $$next-version$$
+ */
+interface Course_Progress_Repository_Interface {
+	/**
+	 * Creates a new course progress.
+	 *
+	 * @param int $course_id The course ID.
+	 * @param int $user_id The user ID.
+	 * @return Course_Progress The course progress.
+	 */
+	public function create( int $course_id, int $user_id ): Course_Progress;
+
+	/**
+	 * Gets a course progress.
+	 *
+	 * @param int $course_id The course ID.
+	 * @param int $user_id The user ID.
+	 * @return Course_Progress|null The course progress.
+	 */
+	public function get( int $course_id, int $user_id ): ?Course_Progress;
+
+	/**
+	 * Checks if a course progress exists.
+	 *
+	 * @param int $course_id The course ID.
+	 * @param int $user_id The user ID.
+	 * @return bool Whether the course progress exists.
+	 */
+	public function has( int $course_id, int $user_id ): bool;
+
+	/**
+	 * Save course progress.
+	 *
+	 * @param Course_Progress $course_progress The course progress.
+	 */
+	public function save( Course_Progress $course_progress ): void;
+}
