--- includes/student-progress/lesson-progress/repositories/class-lesson-progress-repository-factory.php
@@ -1,28 +1,0 @@
-<?php
-/**
- * File containing the Lesson_Progress_Repository_Factory class.
- *
- * @package sensei
- */
-
-namespace Sensei\Student_Progress\Lesson_Progress\Repositories;
-
-if ( ! defined( 'ABSPATH' ) ) {
-	exit;
-}
-
-/**
- * Class Lesson_Progress_Repository_Factory.
- *
- * @since $$next-version$$
- */
-class Lesson_Progress_Repository_Factory {
-	/**
-	 * Creates a new lesson progress repository.
-	 *
-	 * @return Lesson_Progress_Repository_Interface The repository.
-	 */
-	public function create(): Lesson_Progress_Repository_Interface {
-		return new Comments_Based_Lesson_Progress_Repository();
-	}
-}
