--- tests/unit-tests/student-progress/lesson-progress/repositories/test-class-comments-based-lesson-progress-repository.php
@@ -1,150 +1,0 @@
-<?php
-
-namespace SenseiTest\Student_Progress\Lesson_Progress\Repositories;
-
-use Sensei\Student_Progress\Lesson_Progress\Models\Lesson_Progress;
-use Sensei\Student_Progress\Lesson_Progress\Repositories\Comments_Based_Lesson_Progress_Repository;
-
-/**
- * Tests for the Comments_Based_Lesson_Progress_Repository_Test class.
- *
- * @covers \Sensei\Student_Progress\Lesson_Progress\Repositories\Comments_Based_Lesson_Progress_Repository
- */
-class Comments_Based_Lesson_Progress_Repository_Test extends \WP_UnitTestCase {
-	protected $factory;
-
-	public function setup() {
-		parent::setup();
-		$this->factory = new \Sensei_Factory();
-	}
-
-	public function tearDown() {
-		parent::tearDown();
-		$this->factory->tearDown();
-	}
-
-	public function testGet_WhenStatusFound_ReturnsLessonProgress(): void {
-		/* Arrange. */
-		$lesson_id  = $this->factory->lesson->create();
-		$user_id    = $this->factory->user->create();
-		$repository = new Comments_Based_Lesson_Progress_Repository();
-		\Sensei_Utils::update_lesson_status( $user_id, $lesson_id, 'in-progress' );
-
-		/* Act. */
-		$progress = $repository->get( $lesson_id, $user_id );
-
-		/* Assert. */
-		$expected = [
-			'user_id'   => $user_id,
-			'lesson_id' => $lesson_id,
-			'status'    => 'in-progress',
-		];
-		self::assertSame( $expected, $this->export_progress( $progress ) );
-	}
-
-	public function testGet_WhenStatusNotFound_ReturnsNull(): void {
-		/* Arrange. */
-		$lesson_id  = $this->factory->lesson->create();
-		$user_id    = $this->factory->user->create();
-		$repository = new Comments_Based_Lesson_Progress_Repository();
-
-		/* Act. */
-		$progress = $repository->get( $lesson_id, $user_id );
-
-		/* Assert. */
-		self::assertNull( $progress );
-	}
-
-	public function testGet_WhenCreated_ReturnsSameProgress(): void {
-		/* Arrange. */
-		$lesson_id  = $this->factory->lesson->create();
-		$user_id    = $this->factory->user->create();
-		$repository = new Comments_Based_Lesson_Progress_Repository();
-		$created    = $repository->create( $lesson_id, $user_id );
-
-		/* Act. */
-		$progress = $repository->get( $lesson_id, $user_id );
-
-		/* Assert. */
-		self::assertSame( $this->export_progress( $created ), $this->export_progress( $progress ) );
-	}
-
-	public function testHas_WhenStatusFound_ReturnsTrue(): void {
-		/* Arrange. */
-		$lesson_id  = $this->factory->lesson->create();
-		$user_id    = $this->factory->user->create();
-		$repository = new Comments_Based_Lesson_Progress_Repository();
-		\Sensei_Utils::update_lesson_status( $user_id, $lesson_id, 'in-progress' );
-
-		/* Act. */
-		$has = $repository->has( $lesson_id, $user_id );
-
-		/* Assert. */
-		self::assertTrue( $has );
-	}
-
-	public function testHas_WhenStatusNotFound_ReturnsFalse(): void {
-		/* Arrange. */
-		$lesson_id  = $this->factory->lesson->create();
-		$user_id    = $this->factory->user->create();
-		$repository = new Comments_Based_Lesson_Progress_Repository();
-
-		/* Act. */
-		$has = $repository->has( $lesson_id, $user_id );
-
-		/* Assert. */
-		self::assertFalse( $has );
-	}
-
-	public function testGet_WhenProgressChangedAndSaved_ReturnsUpdatedProgress(): void {
-		/* Arrange. */
-		$lesson_id  = $this->factory->lesson->create();
-		$user_id    = $this->factory->user->create();
-		$repository = new Comments_Based_Lesson_Progress_Repository();
-		$progress   = $repository->create( $lesson_id, $user_id );
-		$progress->complete();
-		$repository->save( $progress );
-
-		/* Act. */
-		$actual = $repository->get( $lesson_id, $user_id );
-
-		/* Assert. */
-		self::assertSame( $this->export_progress( $progress ), $this->export_progress( $actual ) );
-	}
-
-	public function testCount_WhenNoProgress_ReturnsZero(): void {
-		/* Arrange. */
-		$course_id  = $this->factory->course->create();
-		$user_id    = $this->factory->user->create();
-		$repository = new Comments_Based_Lesson_Progress_Repository();
-
-		/* Act. */
-		$count = $repository->count( $course_id, $user_id );
-
-		/* Assert. */
-		self::assertSame( 0, $count );
-	}
-
-	public function testCount_WhenProgressCreated_ReturnsOne(): void {
-		/* Arrange. */
-		$course_id  = $this->factory->course->create();
-		$lesson_id  = $this->factory->lesson->create( [ 'meta_input' => [ '_lesson_course' => $course_id ] ] );
-		$user_id    = $this->factory->user->create();
-		$repository = new Comments_Based_Lesson_Progress_Repository();
-		$repository->create( $lesson_id, $user_id );
-
-		/* Act. */
-		$count = $repository->count( $course_id, $user_id );
-
-		/* Assert. */
-		self::assertSame( 1, $count );
-	}
-
-	private function export_progress( Lesson_Progress $progress ): array {
-		return [
-			'user_id'   => $progress->get_user_id(),
-			'lesson_id' => $progress->get_lesson_id(),
-			'status'    => $progress->get_status(),
-		];
-	}
-}
